fix potential overflow issue

+++ kalman_tracking_live.py
@@ -69,8 +69,8 @@ def on_mouse(event, x, y, flags, params):
 # return centre of a set of points representing a rectangle
 
 def center(points):
-    x = (points[0][0] + points[1][0] + points[2][0] + points[3][0]) / 4.0
-    y = (points[0][1] + points[1][1] + points[2][1] + points[3][1]) / 4.0
+    x = np.float32((points[0][0] + points[1][0] + points[2][0] + points[3][0]) / 4.0)
+    y = np.float32((points[0][1] + points[1][1] + points[2][1] + points[3][1]) / 4.0)
     return np.array([np.float32(x), np.float32(y)], np.float32)
 
 #####################################################################
